Fix @Patch and @Put endpoints into ProductController. And add Config Class. The Config Class Allow to create a Baan for the ObjectMapper.

+++ src/main/java/ken/tar/Product/controller/ProductController.java
@@ -8,6 +8,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/products")
@@ -29,41 +30,59 @@ public class ProductController {
     @GetMapping("/{id}")
     public Product getProduct(@PathVariable Long id){
         Product theProduct = theProductService.getProduct(id);
+        if (theProduct == null) {
+            throw new RuntimeException("Product not found with id: " + id);
+        }
         return theProduct;
     }
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping
     public Product saveProduct(Product product){
+        // also just in case they pass an id in JSON ... set id to 0
+        // this is to force a save of new item ... instead of update
+        product.setId(0);
         return theProductService.saveProduct(product);
     }
 
     @ResponseStatus(HttpStatus.ACCEPTED)
     @PutMapping("/{id}")
     public Product updateProduct(@PathVariable Long id, @RequestBody Product product) {
-        Product existingProduct = theProductService.getProduct(id);
-        if (existingProduct == null) {
-            throw new RuntimeException("Product not found with id: " + id);
+        Product dbProduct = theProductService.getProduct(id);
+
+        if (dbProduct == null) {
+            throw new RuntimeException("Product id in path and request body do not match");
         }
-        existingProduct.setName(product.getName());
-        existingProduct.setPrice(product.getPrice());
-        return theProductService.saveProduct(existingProduct);
+
+        dbProduct.setName(product.getName());
+        dbProduct.setPrice(product.getPrice());
+        return theProductService.saveProduct(dbProduct);
     }
 
     @ResponseStatus(HttpStatus.ACCEPTED)
     @PatchMapping("/{id}")
-    public Product patchProduct(@PathVariable Long id, @RequestBody Product product) {
+    public Product patchProduct(@PathVariable Long id, @RequestBody Map<String, Object> patchData) {
         Product existingProduct = theProductService.getProduct(id);
         if (existingProduct == null) {
             throw new RuntimeException("Product not found with id: " + id);
         }
 
-        return theProductService.saveProduct(existingProduct);
+        if (patchData.containsKey("id")) {
+            throw  new RuntimeException("Product id is not allowed in patch request");
+        }
+
+        Product patchedProduct = theProductService.patch(patchData, existingProduct);
+
+        return theProductService.saveProduct(patchedProduct);
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
     public void deleteProduct(@PathVariable Long id){
+        Product theProduct = theProductService.getProduct(id);
+        if (theProduct == null) {
+            throw new RuntimeException("Product not found with id: " + id);
+        }
         theProductService.deleteProduct(id);
     }
 }
